node/requirements.txt
Werkzeug==0.9.6
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
docker-py==0.5.0
ecdsa==0.11
ipdb==0.8
ipython==2.2.0
paramiko==1.15.1
pycrypto==2.6.1
requests==2.4.1
scp==0.8.0
six==1.8.0
websocket-client==0.18.0
wsgiref==0.1.2
